Started writing order logic.

+++ backend/ecommerce/src/main/java/com/sanskar/ecommerce/dto/order/OrderItemRequest.java
@@ -1,6 +1,6 @@
 package com.sanskar.ecommerce.dto.order;
 
 public record OrderItemRequest(
-        int productId,
+        Long productId,
         int quantity
 ) {}
